Get the stack project structure working
* Updated the cabal file with dependencies
* Moved components to better named files
* Changed enough to get compiling
* Updated Readme of example of how to build and run
* Add a single new test file
