+++ src/Main.hs
@@ -2,28 +2,13 @@
 
 import System.Exit
 import Text.Megaparsec
-import VariationalCompiler
+import VariationalCompiler.Parser
+import VariationalCompiler.Json
 import Data.Aeson
 import Data.ByteString.Lazy.Char8(ByteString, putStrLn)
 import Prelude hiding (putStrLn)
 
 -- Aeson instances for Megaparsec
-
-instance ToJSON SourcePos where
-  toJSON pos = object [
-      "sourceName" .= sourceName pos
-    , "sourceLine" .= sourceLine pos
-    , "sourceColumn" .= sourceColumn pos
-    ]
-
-instance ToJSON Message where
-  toJSON = toJSON . messageString
-
-instance ToJSON ParseError where
-  toJSON err = object [
-      "errorPos" .= errorPos err,
-      "errorMessages" .= errorMessages err
-    ]
 
 -- | Allows for custom encoding behavior of Either values.
 newtype Either' a b = E (Either a b)
