+++ testfiles/hello.vjava
@@ -1,0 +1,12 @@
+public class HelloWorld {
+
+    public static void main(String[] args) {
+        // Prints "Hello, World" to the terminal window.
+        #dimension HI
+        System.out.println("Hi");
+        #else
+        System.out.println("Hello, World");
+        #end
+    }
+
+}
