+++ variational-compiler.cabal
@@ -1,19 +1,29 @@
 name:                variational-compiler
 version:             0.1.0.0
 synopsis:            Compiler for text files with variational constructs
-description:         Please see README.md
-homepage:            https://github.com/rikkigibson/variational-compiler#readme
+homepage:            https://github.com/rikkigibson/variational-compiler
 license:             MIT
-license-file:        LICENSE
 author:              Rikki Gibson and Cody Ray Hoeft
 maintainer:          rikkigibson@gmail.com
-copyright:           2016 Rikki Gibson and Cody Ray Hoeft
-category:            Web
+category:            Language
 build-type:          Simple
-cabal-version:       >=1.10
+cabal-version:       >= 1.22
+
+source-repository head
+  type:     git
+  location: https://github.com/rikkigibson/variational-compiler
+  branch:   master
 
 executable variational-compiler
   hs-source-dirs:      src
   main-is:             Main.hs
   default-language:    Haskell2010
-  build-depends:       base >= 4.7 && < 5
+  build-depends:       
+    base,
+    aeson,
+    megaparsec,
+    bytestring,
+    scientific,
+    text,
+    aeson-pretty
+  -- ghc-options:
